put&delete users, delete projects

+++ project-ms/app/router.py
@@ -1,58 +1,50 @@
-from flask import request
-from flask import Blueprint
+from flask import request, Blueprint
 from app.controllers.project_controller import ProjectController
 from app.controllers.image_controller import ImageController
 from app.controllers.tool_controller import ToolController
 app_router = Blueprint('app_router', __name__)
 
-# Routes for project management
-@app_router.route('/projects', methods=['GET', 'POST'])
+# Routes for projects management
+@app_router.route('/projects', methods=['GET', 'POST', 'DELETE'])
 def projects():
     """
     GET: Retrieve all projects for the current user.
     POST: Create a new project for the current user.
+    DELETE: Delete all projects of the current user.
     """
-    if request.method == 'POST':
+    if request.method == 'DELETE':
+        return ProjectController.delete_projects_user()
+    elif request.method == 'POST':
         return ProjectController.add_project()
     else:  # Default to GET for fetching projects
         return ProjectController.get_projects_by_user()
 
-@app_router.route('/projects/<project_id>', methods=['PUT'])
-def update_project(project_id):
-    """
-    PUT: Update the name of a specific project (owned by the user).
-    """
-    return ProjectController.update_project(project_id)
-
-@app_router.route('/projects/<project_id>', methods=['GET'])
-def get_project_by_id(project_id):
+# Routes for a specific project management
+@app_router.route('/projects/<project_id>', methods=['GET', 'PUT', 'DELETE'])
+def project(project_id):
     """
     GET: Retrieve details of a specific project by its ID.
-    """
-    return ProjectController.get_project_by_id(project_id)
-
-
-@app_router.route('/projects/<project_id>', methods=['DELETE'])
-def delete_project(project_id):
-    """
+    PUT: Update the name of a specific project (owned by the user).
     DELETE: Delete a specific project (owned by the user).
     """
-    return ProjectController.delete_project(project_id)
-
+    if request.method == 'DELETE':
+        return ProjectController.delete_project(project_id)
+    elif request.method == 'PUT':
+        return ProjectController.update_project(project_id)
+    else:  # Default to GET for fetching project details
+        return ProjectController.get_project_by_id(project_id)
+    
 # Routes for image management
-@app_router.route('/projects/<project_id>/images', methods=['POST'])
-def upload_image(project_id):
-    """
-    POST: Upload an image to a specific project's source bucket.
-    """
-    return ImageController.upload_image(project_id)
-
-@app_router.route('/projects/<project_id>/images', methods=['GET'])
-def get_images(project_id):
+@app_router.route('/projects/<project_id>/images', methods=['GET', 'POST'])
+def images(project_id):
     """
     GET: Retrieve all processed image links from a project's output bucket.
+    POST: Upload an image to a specific project's source bucket.
     """
-    return ImageController.get_images(project_id)
+    if request.method == 'POST':
+        return ImageController.upload_image(project_id)
+    else:  # Default to GET for fetching images
+        return ImageController.get_images(project_id)
 
 @app_router.route('/projects/<project_id>/images/<image_id>', methods=['DELETE'])
 def delete_image(project_id, image_id):
@@ -62,33 +54,27 @@ def delete_image(project_id, image_id):
     return ImageController.delete_image(project_id, image_id)
 
 # Routes for tool management
-@app_router.route('/projects/<project_id>/tools', methods=['GET'])
-def get_tools(project_id):
+@app_router.route('/projects/<project_id>/tools', methods=['GET', 'POST'])
+def tools(project_id):
     """
     GET: Retrieve all tools associated with a specific project.
-    """
-    return ToolController.get_tools(project_id)
-
-@app_router.route('/projects/<project_id>/tools', methods=['POST'])
-def add_tool(project_id):
-    """
     POST: Add a new tool to a specific project.
     """
-    return ToolController.add_tool(project_id)
+    if request.method == 'POST':
+        return ToolController.add_tool(project_id)
+    else:  # Default to GET for fetching tools
+        return ToolController.get_tools(project_id)
 
-@app_router.route('/projects/<project_id>/tools/<tool_id>', methods=['PUT'])
-def update_tool(project_id, tool_id):
+@app_router.route('/projects/<project_id>/tools/<tool_id>', methods=['PUT', 'DELETE'])
+def tool(project_id, tool_id):
     """
     PUT: Update a specific tool's configuration for a project.
-    """
-    return ToolController.update_tool(project_id, tool_id)
-
-@app_router.route('/projects/<project_id>/tools/<tool_id>', methods=['DELETE'])
-def delete_tool(project_id, tool_id):
-    """
     DELETE: Remove a specific tool from a project.
     """
-    return ToolController.delete_tool(project_id, tool_id)
+    if request.method == 'PUT':
+        return ToolController.update_tool(project_id, tool_id)
+    else: # Default to DELETE for deleting a tool
+        return ToolController.delete_tool(project_id, tool_id)
 
 # Routes for project processing and status
 @app_router.route('/projects/<project_id>/process', methods=['POST'])
@@ -104,7 +90,6 @@ def get_project_status(project_id):
     GET: Retrieve the processing status of a specific project.
     """
     return ProjectController.get_project_status(project_id)
-
 
 # Route to serve images from MinIO through the backend
 @app_router.route('/images/<bucket>/<path:filename>', methods=['GET'])
